requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.0
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
Pygments==2.7.2
PySocks==1.7.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.2.1
wcwidth==0.2.5
